LDA/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
PyMuPDF==1.23.8
requests==2.31.0
python-dotenv==1.0.1
transformers==4.37.2
torch==2.2.0
huggingface-hub==0.20.3
sentencepiece==0.1.99
protobuf==4.25.3

